requirements.txt
# Requirements for fraud detection training and visual pipeline
google-cloud-aiplatform==1.36.0
google-cloud-storage==2.10.0
google-cloud-tasks==2.14.2
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
numpy==1.24.3
kfp==2.4.0